Move mapping to its own package
